Update productos
Agragadas las imagenes de los productos todas de un tamaño de 1000x1000px en formato jpg, corregido tipo de variable en k_idproducto y k_idpedido de Double a Long

+++ src/java/control/PedidoDAO.java
@@ -37,7 +37,7 @@ public class PedidoDAO {
             try (PreparedStatement prepStmt = conexion.prepareStatement(strSQL)) {
                 ResultSet rs = prepStmt.executeQuery();
                 while (rs.next()) {
-                    pedido.getId_pedido_array().add(rs.getDouble(1));
+                    pedido.getId_pedido_array().add(rs.getLong(1));
                     pedido.getEstado_pedido_array().add(rs.getDouble(2));
                     pedido.getFecha_pedido_array().add(rs.getDate(3));
                     pedido.getTotal_pedido_array().add(rs.getDouble(4));
